fix: PostgreSQL 실제 DB 연동 및 Entity 스키마 완전 정렬
- PostgreSQL 실제 외부 서버 연동 (i13d203.p.ssafy.io:5432/stock_db)
- 모든 Entity를 실제 DB 테이블 구조에 완벽 매핑
- Account Entity created_at 필드 제거 (실제 테이블에 존재하지 않음)
- MockTrade Entity stockId JoinColumn 매핑 수정
- User Entity aiId 타입 Long → Integer 변경
- News Entity keywords 필드 제거 및 스키마 간소화
- StockDay Entity newDate 필드 제거
- 모든 Entity 테이블명을 소문자로 변경 (PostgreSQL 규칙 준수)
- Repository, Service, Controller, DTO 타입 일관성 수정
- 컴파일 에러 및 매핑 오류 전체 해결
- DatabaseTestController에 테이블 구조 확인 엔드포인트 추가
- Swagger UI 그룹화 단순화로 API 접근성 개선
- 73개 API 엔드포인트 정상 등록 및 작동 확인
- 실제 DB 연결 상태 및 20개 테이블 구조 검증 완료

+++ BE/src/main/java/com/example/LAGO/controller/StockController.java
@@ -58,7 +58,7 @@ public class StockController {
             @RequestHeader("User-Id") // TODO: JWT 인증으로 변경 예정
             @NotNull(message = "사용자 ID는 필수입니다") 
             @Positive(message = "사용자 ID는 양수여야 합니다") 
-            Integer userId,
+            Long userId,
             
             @Parameter(description = "매수 요청 정보", required = true)
             @RequestBody 
@@ -119,7 +119,7 @@ public class StockController {
             @RequestHeader("User-Id") // TODO: JWT 인증으로 변경 예정
             @NotNull(message = "사용자 ID는 필수입니다") 
             @Positive(message = "사용자 ID는 양수여야 합니다") 
-            Integer userId,
+            Long userId,
             
             @Parameter(description = "매도 요청 정보", required = true)
             @RequestBody 
@@ -180,7 +180,7 @@ public class StockController {
             @RequestHeader("User-Id") // TODO: JWT 인증으로 변경 예정
             @NotNull(message = "사용자 ID는 필수입니다") 
             @Positive(message = "사용자 ID는 양수여야 합니다") 
-            Integer userId,
+            Long userId,
             
             @Parameter(description = "매매 요청 정보", required = true)
             @RequestBody
